requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.6
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn
idna==3.3
importlib-metadata==4.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
Markdown==3.3.6
MarkupSafe==2.0.1
murmurhash==1.0.6
nltk==3.7
numpy==1.22.2
packaging==21.3
pathy==0.6.1
preshed==3.0.6
psycopg2-binary==2.9.3
pydantic==1.8.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz
regex==2022.1.18
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
smart-open==5.2.1
spacy==3.2.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
thinc==8.0.13
tqdm==4.62.3
typer==0.4.0
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
wasabi==0.9.0
zipp==3.7.0
django-environ
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0.tar.gz#egg=en_core_web_sm==3.2.0
